# [Merged by Bors] - Remove unnecessary alternate create_texture path in prepare_asset for Image