@mockersf 2022-11-18
while this works the same, it is potentially slower in the no-mip/no-compression case as the code was taking a lot of shortcuts