@DGriffin91 2022-11-18
@mockersf 
Benchmarked in vtune with 1 million 8x8px images.
Total time spent in `prepare_asset()` for Image:

Current bevy with branch:
Run 1: 21.299s
Run 2: 21.123s
Run 3: 21.305s

This PR:
Run 1: 19.939s
Run 2: 20.066s
Run 3: 20.596s

So this PR is actually faster. It looks like at least part of this is because `image.is_compressed()` and `image.texture_descriptor.format.pixel_size()` are a bit slow.